fixed the image downloader
fixed a bug with the image downloader downloading corrupt jpgs, now it
checks if the image is corrupted, and if so, we'll download the original
image, and downscale it ourselves. it's quite expensive but it's very
rare that tmdb will provide corrupt images so it should be fine. in the
rare case that the original image is corrupted too, it'll just add a
placeholder image cuz there's nothing else we can do :'(