Added movie plays, and fixed high cpu usage with the ratatui_image backend
Added the ability to add more plays to a movie, each play will have a
timestamp and a rating associated with it, but only the latest play will
be used as the user_rating.
Fixed the high cpu usage issue when using the ratatui_image backend. The
current fix reduces the cpu usage at the cost of a LOT more threads, one
per image. still not ideal but i'll have to find a better fix later.